requirements.txt
icalendar
pytz
requests
discord.py